move sbt/ivy artifact stores to allow caching
drone cache plugins can only cache files located in /drone/. both the
sbt and ivy caches are now moved within this dir, while also keeping the
credentials file outside /drone/ so they can't be persisted
accidentally.

to simplify the process, I'm setting $HOME to /drone/ to allow sbt and
ivy to automatically put files in the right locations. symlinks are also
created in the real home directory that point to this new location